fix: HIGH #13 approvalTimeoutMs, #20 shared extractTaggedJSON, #21 negative depth guard, #22 strict score validation, #24 no responses fallback, #25 filter deleted tasks, CRITICAL #3 O_EXCL sentinel, MEDIUM #4 signoff empty turn, #6 parse counter on any success, #8 token field validation

+++ src/tools/workflow/engine.ts
@@ -295,8 +295,8 @@ export function teamWorkflowTool(ctx: PluginContext): ToolDefinition {
             ),
             ...signoffSchemaFields,
             ...humanApprovalSchemaFields,
-            timeout_ms: tool.schema.number().min(1000).optional(),
-            token_budget: tool.schema.number().min(1).optional().describe(
+            timeout_ms: tool.schema.number().int().min(1000).optional(),
+            token_budget: tool.schema.number().int().min(1).optional().describe(
                 "optional token cap; orchestration fails if exceeded",
             ),
             max_retries: tool.schema.number().int().min(0).max(5).optional().describe(
